тест test_class_Vacancy
